+++ packages/components/src/components/stepper-horizontal/stepper-horizontal/stepper-horizontal-utils.ts
@@ -7,9 +7,7 @@ import type { StepperHorizontalItemInternalHTMLProps } from '../stepper-horizont
 export const STEPPER_HORIZONTAL_SIZES = ['small', 'medium'] as const;
 export type StepperHorizontalSize = (typeof STEPPER_HORIZONTAL_SIZES)[number];
 
-/** @deprecated */
-export type StepperHorizontalUpdateEvent = { activeStepIndex: number };
-export type StepperHorizontalUpdateEventDetail = StepperHorizontalUpdateEvent;
+export type StepperHorizontalUpdateEventDetail = { activeStepIndex: number };
 
 export const getIndexOfStepWithStateCurrent = (stepperHorizontalItems: HTMLPStepperHorizontalItemElement[]): number => {
   return stepperHorizontalItems.findIndex((item) => item.state === 'current');