+++ packages/components/src/utils/typography/typography-text-color.ts
@@ -1,14 +1,3 @@
-/** @deprecated */
-export const TYPOGRAPHY_TEXT_COLORS_DEPRECATED = [
-  'brand', // deprecated
-  'default', // deprecated
-  'neutral-contrast-low', // deprecated
-  'neutral-contrast-medium', // deprecated
-  'neutral-contrast-high', // deprecated
-  'notification-neutral', // deprecated
-] as const;
-/** @deprecated */
-export type TypographyTextColorDeprecated = (typeof TYPOGRAPHY_TEXT_COLORS_DEPRECATED)[number];
 export const TYPOGRAPHY_TEXT_COLORS = [
   'primary',
   'contrast-low',